powercap: idle_inject: Add begin/end callbacks
In some cases idle duration and runtime needs to be modified using a
feedback loop. It is also likely that each CPU needs different values
when the idle inject is registered with a cpumask with more than one
CPU.

Also if there is some scenario, if the idle inject is percent is more
than the required based on the feedback, it should skip calling
play_idle_precise().

Add a capability to register a begin and end callback during call to
idle_inject_register(). If they are not NULL, then begin callback is
called before calling play_idle_precise() and end callback is called
after play_idle_precise().

If begin callback is present and returns non 0 value then
play_idle_precise() is not called.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>