tick: Use local_bh_blocked() check for pr_warn only
Call to local_bh_blocked() will always return false when CONFIG_PREEMPT_RT
is not defined. From the description of the function the purpose is
"guard against false positive softirq pending warnings". This was the
case till Linux kernel version 5.17.

After commit 0345691b24c0 ("tick/rcu: Stop allowing RCU_SOFTIRQ in idle"),
if local_bh_blocked() return false, the function report_idle_softirq()
will always return false. This means tick can be stopped
unconditionally.

If is the real intention, then this check should be moved to the top
of the function and avoid all other checks when CONFIG_PREEMPT_RT
is not defined.

Here the behavior is restored as the kernel version 5.17 and prior
and use only for printing warning.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>